primitives: rename consensus into coding module
